requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
pydub>=0.25.1

# For local TTS (optional - use --engine piper)
# Note: piper-tts 1.3.0 works with Python 3.13, newer versions may have issues
piper-tts==1.3.0
